initial refactor of gql
