+++ api/src/services/specifications.ts
@@ -227,24 +227,24 @@ class OASSpecsService implements SpecificationSubService {
 									requestBody: ['get', 'delete'].includes(method)
 										? undefined
 										: {
-												content: {
-													'application/json': {
-														schema: {
-															oneOf: [
-																{
-																	type: 'array',
-																	items: {
-																		$ref: `#/components/schemas/${tag.name}`,
-																	},
-																},
-																{
+											content: {
+												'application/json': {
+													schema: {
+														oneOf: [
+															{
+																type: 'array',
+																items: {
 																	$ref: `#/components/schemas/${tag.name}`,
 																},
-															],
-														},
+															},
+															{
+																$ref: `#/components/schemas/${tag.name}`,
+															},
+														],
 													},
 												},
-										  },
+											},
+										},
 									responses: {
 										'200': {
 											description: 'Successful request',
@@ -252,18 +252,18 @@ class OASSpecsService implements SpecificationSubService {
 												method === 'delete'
 													? undefined
 													: {
-															'application/json': {
-																schema: {
-																	properties: {
-																		data: {
-																			items: {
-																				$ref: `#/components/schemas/${tag.name}`,
-																			},
+														'application/json': {
+															schema: {
+																properties: {
+																	data: {
+																		items: {
+																			$ref: `#/components/schemas/${tag.name}`,
 																		},
 																	},
 																},
 															},
-													  },
+														},
+													},
 										},
 									},
 								},
@@ -285,30 +285,30 @@ class OASSpecsService implements SpecificationSubService {
 									requestBody: ['get', 'delete'].includes(method)
 										? undefined
 										: {
-												content: {
-													'application/json': {
-														schema: {
-															$ref: `#/components/schemas/${tag.name}`,
-														},
+											content: {
+												'application/json': {
+													schema: {
+														$ref: `#/components/schemas/${tag.name}`,
 													},
 												},
-										  },
+											},
+										},
 									responses: {
 										'200': {
 											content:
 												method === 'delete'
 													? undefined
 													: {
-															'application/json': {
-																schema: {
-																	properties: {
-																		data: {
-																			$ref: `#/components/schemas/${tag.name}`,
-																		},
+														'application/json': {
+															schema: {
+																properties: {
+																	data: {
+																		$ref: `#/components/schemas/${tag.name}`,
 																	},
 																},
 															},
-													  },
+														},
+													},
 										},
 									},
 								},
